Explicity initialize Archive type at class creation time
The GobMemoryArchive does not define an ARCHIVE_GOB type anywhere.
To fix this, add a new constructor to the base class to set the
archive type at instantiation time.
This fixes a crash on Linux with Mods (quitting while running a
mod, or loading a savegame away from a mod, where the memory gob
freeing causes a crash).

Signed-off-by: Manuel Lauss <manuel.lauss@gmail.com>