Linux: default linkage to legacy OpenGL (libGL.so) library
Modern CMake (and when I initially wrote the CMakeLists) prefers
GLVND, which is a fancy way to dispatch OpenGL calls over multiple
different client libraries.  However that does not work with the
ancient (<=340.xx) nvidia binary linux drivers, which still assume
that there can be only one libGL.so on a linux system.

Even if the GLVND libraries are available on the target system,
there will not be any GL output since the client library does
not understand the GLVND dispatches, resulting in black screens.

Fix this by explicitly requesting CMake to prefer the legacy
OpenGL linkage approach.

Fixes #395

Signed-off-by: Manuel Lauss <manuel.lauss@gmail.com>