IMuse: handle type 2 chunks
In Dark Tide 1, there's the MCHN3.VOC file which contains a
voc chunk with id 2, which causes TFE to spam the log with
an unending stream of:

[Error : iMuse] ERR: Illegal chunk in sound    2561ba3c54df0...
[Error : iMuse] ERR: Illegal chunk in sound    3561ba3cb5658...
[Error : iMuse] ERR: Illegal chunk in sound    5561ba3cb5658...

Fix this by handling type 2 chunks.

Also clarify the debug output a bit and stop spamming the log
on known but unhandled chunks (3 and 5).