* Updates Windows SDL to version 2.30.6. * Fixes thread contention between the midi and audio threads when synthesizing midi audio. * Report an error if SDL_SetRelativeMouseMode() fails and allow it to retry again on future frames (due to buggy mouse drivers).
After updating SDL I noticed that the midi was garbled sounded in some cases
and discovered that the problem "went away" if I removed the lock in synthesizeMidi()
and after looking at it further realized there can be content between the midi
and audio thread mutexes in this case.

However just removing the lock would bring back crashes when changing midi
device, so I split the mutex into two - the midi thread mutex, and the device
change mutex - which allows me to freely change devices/outputs but remove
the contention.