Cutscene LFD handling refinements
Mount the Cutscene LFDs on-demand and not all of them at once during
game startup:  The issue is that some of the cutscene LFDs contain
identically named DELT and ANIM files but with different content.
This could especially be seen in the cutscene when launching the
Detention Center level, in the scnene where Mohc and Vader are talking:
The Body of Mohc is completely wrong, probably sourced from another
LFD.
Since the cutscene list also names the LFDs to use, the LFDs are
mounted and unmonted on demand, restoring essentially the logic of
the main branch.

Add more debug output to vpXX functions

Fix up FrontEndUi game source path/archive handling some more
(still not fully complete)